Removing vitals dependencies.
This could not be done in several commits easly. We will rather start testing on another branch and eventually merge the things if it is stable.
Here is the scope of the changes:
- Changing the observer to atlas implementation
- Changing thread to atlas implementation
- Removing Unit Test from vitals and create Google Unit Test files on vision_lib.
- Removing the pointers object and go for shared_ptr everywhere.
- Fixes due to all theses changes.

The vision_server is not feeding the video at this point, going to fix it in another commit.