Fixing file not opening issue.
Videos are open at construction, and there was a execption throw when trying to open a video already open, which
stopped the video opening. Remove this call, since if it is already open, then its just less code to run...

Removed the file opening in file_context because a video file is open upon  construction.