Enabling dual camera streaming at low resolution.
The camera can now stream a the same time. HOWEVER
we have 640*480 images. Next dev will try to fix that.

Probably something to do with iso-channels. In Coriander, when
both device are streaming, there is one bus at 53% of its bandwith.
when streaming at 800*600, it does not work. Individually, it takes
42% for each cam(probably something about
the 80%async/20%sync thing...)

Will try to do two bus for streaming.